requirements.txt
dash>=2.5.1
numpy>=1.22.4
opencv-python>=4.4.0
pandas>=2.0.0
Pillow>=10.0.0
plotly>=5.8.2
torch>=2.0.0
torchvision>=0.15.0
parse
tqdm
